requirements.txt
streamlit>=1.36.0
pydantic>=2.8.2
python-docx>=1.1.2
pypdf>=4.3.1
openai>=1.45.0
tiktoken>=0.7.0
transformers>=4.42.4
torch>=2.3.1 ; platform_system != "Darwin" or platform_machine != "arm64"
numpy>=1.26.4
pillow 
pytesseract 
PyPDF2 
langchain 
langchain-community 
faiss-cpu 
sentence-transformers 
google-cloud-aiplatform
langchain_google_vertexai
tesseract
